--- src/backend/internal/controller/photographer/req_refund_booking.go
@@ -1,47 +1,0 @@
-package photographer
-
-import (
-	"net/http"
-
-	"github.com/Roongkun/software-eng-ii/internal/controller/util"
-	"github.com/Roongkun/software-eng-ii/internal/model"
-	"github.com/gin-gonic/gin"
-	"github.com/google/uuid"
-)
-
-func (r *Resolver) RequestRefundBooking(c *gin.Context) {
-	photographer, ok := r.getPhotographer(c)
-	if !ok {
-		return
-	}
-
-	paramId := c.Param("id")
-	bookingId := uuid.MustParse(paramId)
-
-	booking, err := r.BookingUsecase.BookingRepo.FindOneById(c, bookingId)
-	if err != nil {
-		util.Raise500Error(c, err)
-		return
-	}
-	gallery, err := r.GalleryUsecase.GalleryRepo.FindOneById(c, booking.GalleryId)
-	if err != nil {
-		util.Raise500Error(c, err)
-		return
-	}
-
-	if gallery.PhotographerId != photographer.Id {
-		util.Raise403Error(c, "this is not your booking")
-		return
-	}
-
-	booking.Status = model.BookingRefundReqStatus
-	if err := r.BookingUsecase.BookingRepo.UpdateOne(c, booking); err != nil {
-		util.Raise500Error(c, err)
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"status": "success",
-		"data":   booking,
-	})
-}
