+++ src/backend/internal/controller/user/extract_methods.go
@@ -1,0 +1,18 @@
+package user
+
+import (
+	"github.com/Roongkun/software-eng-ii/internal/controller/util"
+	"github.com/Roongkun/software-eng-ii/internal/model"
+	"github.com/gin-gonic/gin"
+)
+
+func getUser(c *gin.Context) (*model.User, bool) {
+	user := c.MustGet("user")
+	userObj, ok := user.(model.User)
+	if !ok {
+		util.Raise400Error(c, "could not bind json")
+		return nil, false
+	}
+
+	return &userObj, true
+}
